+++ pipeline/__init__.py
@@ -1,6 +1,5 @@
 from .pipeline import SupervisedPipeline
-from .lr_schedulers import PolynomialDecay
 
 __name__ = [
-    'SupervisedPipeline', 'PolynomialDecay'
+    'SupervisedPipeline'
 ]
